DB鯖からの投稿削除に対応

+++ Dialog.js
@@ -9,21 +9,42 @@ window.addEventListener("DOMContentLoaded", () => {
 				if (event.ctrlKey && event.keyCode == 13) dialog.querySelector('Button[Data-Action="Dialog_Submit"]').click();
 			});
 		}
-	});
-	
-	DOM('@Dialog Button[Data-Action="Dialog_Close"]').forEach((btn) => {
-		btn.addEventListener("click", () => {
-			btn.offsetParent.close();
+
+		dialog.querySelectorAll('Dialog *[Required]').forEach((input) => {
+			input.addEventListener("input", () => {
+				let result = true;
+
+				dialog.querySelectorAll('Dialog *[Required]').forEach(requiredField => {
+					if (requiredField.value.replace(/\s/g, "").length == 0) {
+						result = false;
+						return;
+					}
+				});
+
+				if (result) {
+					dialog.querySelector('Button[Data-Action="Dialog_Submit"]').classList.remove("mdl-button--disabled");
+				} else {
+					dialog.querySelector('Button[Data-Action="Dialog_Submit"]').classList.add("mdl-button--disabled");
+				}
+			});
+		});
+		
+		dialog.querySelectorAll('Dialog Button[Data-Action="Dialog_Close"]').forEach((btn) => {
+			btn.addEventListener("click", () => {
+				btn.offsetParent.close();
+			});
 		});
 	});
 
 
 
-	DOM("#Dialogs_Profile_DeleteConfirmer_Btns_Yes").addEventListener("click", () => {
-		if (DOM("#Dialogs_Profile_DeleteConfirmer_Content_Email_Input").value == base.user.email) {
-			base.delete();
-		} else {
-			DOM("#Dialogs_Profile_DeleteConfirmer_Content_Email").classList.add("is-invalid");
+	DOM("#Dialogs_Profile_DeleteConfirmer_Btns_Yes").addEventListener("click", (event) => {
+		if (!event.currentTarget.classList.contains("mdl-button--disabled")) {
+			if (DOM("#Dialogs_Profile_DeleteConfirmer_Content_Email_Input").value == base.user.email) {
+				base.delete();
+			} else {
+				DOM("#Dialogs_Profile_DeleteConfirmer_Content_Email").classList.add("is-invalid");
+			}
 		}
 	});
 
@@ -61,72 +82,54 @@ window.addEventListener("DOMContentLoaded", () => {
 
 
 
-	[DOM("#Dialogs_Thread_InfoInputer_Content_Name_Input"), DOM("#Dialogs_Thread_InfoInputer_Content_Overview_Input"), DOM("#Dialogs_Thread_InfoInputer_Content_Detail_Input")].forEach((textarea) => {
-		textarea.addEventListener("input", (event) => {
-			if (event.target.value.replace(/\s/g, "").length == 0) {
-				DOM("#Dialogs_Thread_InfoInputer_Btns_OK").classList.add("mdl-button--disabled");
-			} else {
-				DOM("#Dialogs_Thread_InfoInputer_Btns_OK").classList.remove("mdl-button--disabled");
-			}
-		});
-	});
+	DOM("#Dialogs_Thread_InfoInputer_Btns_OK").addEventListener("click", (event) => {
+		if (!event.currentTarget.classList.contains("mdl-button--disabled")) {
+			base.Database.transaction("threads", (res) => {
+				let now = new Date().getTime();
 
-	DOM("#Dialogs_Thread_InfoInputer_Btns_OK").addEventListener("click", () => {
-		base.Database.transaction("threads", (res) => {
-			let now = new Date().getTime();
+				base.Database.set("threads/" + res.length, {
+					title: DOM("#Dialogs_Thread_InfoInputer_Content_Name_Input").value,
+					overview: DOM("#Dialogs_Thread_InfoInputer_Content_Overview_Input").value,
+					detail: DOM("#Dialogs_Thread_InfoInputer_Content_Detail_Input").value,
 
-			base.Database.set("threads/" + res.length, {
-				title: DOM("#Dialogs_Thread_InfoInputer_Content_Name_Input").value,
-				overview: DOM("#Dialogs_Thread_InfoInputer_Content_Overview_Input").value,
-				detail: DOM("#Dialogs_Thread_InfoInputer_Content_Detail_Input").value,
+					jobs: {
+						Owner: (() => {
+							let owner = {}; owner[base.user.uid] = "";
+							return owner;
+						})(),
 
-				jobs: {
-					Owner: (() => {
-						let owner = {}; owner[base.user.uid] = "";
-						return owner;
-					})(),
+						Admin: {
 
-					Admin: {
+						}
+					},
 
-					}
-				},
+					createdAt: now,
+					
+					data: [
+						{
+							uid: base.user.uid,
+							content: DOM("#Dialogs_Thread_InfoInputer_Content_Name_Input").value,
+							plusCount: 0,
+							createdAt: now
+						}
+					],
 
-				createdAt: now,
+					password: ""
+				});
 				
-				data: [
-					{
-						uid: base.user.uid,
-						content: DOM("#Dialogs_Thread_InfoInputer_Content_Name_Input").value,
-						plusCount: 0,
-						createdAt: now
-					}
-				],
-
-				password: ""
+				DOM("#Dialogs_Thread_InfoInputer").close();
+				parent.document.querySelector("IFrame.mdl-layout__content").src = "Thread/Viewer/?tid=" + res.length;
 			});
-			
-			DOM("#Dialogs_Thread_InfoInputer").close();
-			parent.document.querySelector("IFrame.mdl-layout__content").src = "Thread/Viewer/?tid=" + res.length;
-		});
-	});
-
-
-
-	DOM("#Dialogs_Thread_Poster_Content_Value_Input").addEventListener("input", (event) => {
-		if (event.target.value.replace(/\s/g, "").length == 0) {
-			DOM("#Dialogs_Thread_Poster_Btns_OK").classList.add("mdl-button--disabled");
-		} else {
-			DOM("#Dialogs_Thread_Poster_Btns_OK").classList.remove("mdl-button--disabled");
 		}
 	});
 
+	
+
 	DOM("#Dialogs_Thread_Poster_Btns_OK").addEventListener("click", (event) => {
-		if (!DOM("#Dialogs_Thread_Poster_Btns_OK").classList.contains("mdl-button--disabled")) {
+		if (!event.currentTarget.classList.contains("mdl-button--disabled")) {
 			DOM("#Screens_Loading").removeAttribute("Disabled");
 
 			base.Database.transaction("threads/" + DOM("#Dialogs_Thread_Poster_Content_TID").value + "/data", (res) => {
-				console.log(res);
-
 				base.Database.set("threads/" + DOM("#Dialogs_Thread_Poster_Content_TID").value + "/data/" + res.length, {
 					uid: base.user.uid,
 					content: DOM("#Dialogs_Thread_Poster_Content_Value_Input").value,
